Added rect_intersect
A simple function to check whether two rectangles are intersecting. This
is done by seeing if the x co-ordinate of rect2 liest between the x
co-ordinates of the top-left and bottom-right points of rect1. If so
then it checks for the same in y co-ordinates and returns the correct
boolean.